add service interface and classes template

+++ lab2/src/main/java/com/cmpe275/lab2/dao/PlayerRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.cmpe275.lab2.model.Player;
 
-public interface PlayerRepository extends CrudRepository<Player, Integer> {
+public interface PlayerRepository extends CrudRepository<Player, Long> {
 	
 }
